[testing] Add "CRASH" summary.json status
The mobly test "ERROR" status can map to the resultdb "CRASH" status to
indicate that the test crashed in some way that may not be the test's
fault. This may also be useful in the future for tests that fail due to
infrastructure reasons, e.g. the target becomes unresponsive.

The primary consumers of summary.json (recipes and Sponge processor) are
already resilient to new test statuses being added; any test with an
unknown status will be considered to have failed, which is fine in this
case.

Fixed: 116847
Change-Id: I1e74b10baf75807d378a329cc876d4a340f6544f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773164
Reviewed-by: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Cho <chok@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>