[netstack3] Implement SO_BINDTODEVICE for TCP
Start using the unused device field in the TCP socket map key and allow
setting it for unbound, bound, listener, and connected sockets, and
retrieving the value. Implement the same unroutable check as for UDP
sockets, where a connected socket cannot be bound to a device if that
would make the socket unroutable.

Bug: 102103
Change-Id: Ie87d72b3e0067ed65ebdabe1a75f51c3eed139b3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729942
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>