[run-boot-test] Fix matching bug and clarify build requirements
This change updates `fx run-boot-test` to state its build requirements
more helpfully and also to update the fuzzy matching logic to be more
helpful. In particular, we fix a previous bug that would view an exact
name as ambiguous if it appeared as a substring in another name.

Change-Id: I8a45ea7362e00a46862c773bd731d487a267cce0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768770
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>