[fidl][fix] Change "fix" to "fixable"
We'll use the term "fixable" to name the description of something we can
automatically change, and "fix" to name the actual class that does the
transformation specified by a given fixable.

I've also expanded the information associated with a fixable: it now
tracks required experimental flags, as well as the "scope", or depth of
information needed to perform the fix.

Test: fx test fidlc-test
Bug: 114357
Change-Id: Ia491835589644b162b64fc2880b9246061e1a785
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/772222
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Alex Zaslavsky <azaslavsky@google.com>