[ffx_trace] Add category group support to ffx trace
The goal of category groups is to make it easier to work with
large or unfamiliar lists of trace categories. It allows maintainers
to add predefined curated tracing groups that can be used by passing
the name prefixed with # in the category list provided to ffx trace.

The implementation uses ffx config to store the information about
category groups. New groups can be added to the global config by editing
data/config.json in the trace plugin.

Bug: 115758

Change-Id: Iea576ae64ec7a4d75e208055bbd39e85077f4c75
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761184
Reviewed-by: Kevin Marshall <marshallk@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Bryant Chandler <bryantchandler@google.com>