[assembly] Enable account partition using separate assembly input
The GN flag `include_account_in_fvm` will be used to generate a separate
account config to be passed to assembly. When this config is not
supplied, the account will not be added.

This change allows us to declare explicit GN targets for all the images
config, and point to them in the board definitions.

Bug: 116083
Change-Id: I27cda27c91281bb7c74705d1a7e6946487cbe727
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773723
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Iserovich <diserovich@google.com>
Reviewed-by: Daniel Young <youngda@google.com>