[trace] Acquire a Context when Checking if Enabled
When we enabled tracing, we
1) Enabled the tracing state
2) Set up the context
3) Enabled getting a context

This means if a trace provider quickly started tracing after observing
g_state being started, it could attempt to grab the trace an event,
which would try to grab the context, find that it's empty, and then bail
silently.

This changes trace_is_enabled to check that it's able to get a context
before indicating that the trace has started. So that callers don't try
to quickly emit a trace event before the system is actually ready.

Fixed: 114682
Change-Id: Ibbfaaf1a7903ed054ad5dfc889abc0af32abf068
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762169
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Reviewed-by: Pierre Labatut <plabatut@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
API-Review: Christopher Johnson <crjohns@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>