[kernel][vm] Convert the PageQueues lock to a spinlock.
Change the lock currently used by the implementation of the VM's
PageQueues lock from a CriticalMutex to Spinlock.

This lock currently protects most of the state of the PageQueues
object.  In general, there should be one of these objects per NUMA
node in the system, but in practice there is only one since we do not
currently run on many (any?) systems with multiple NUMA nodes.

Instrumentation has shown that, generally speaking, the amount of work
being done inside of this lock is quite low.  Typically, on the order
of 300nSec or so in a nested virtualised environment (something of a
worst case scenario).  When the lock is not being heavily contested,
this is fine.  It only takes a handful of nanoseconds (20 or so) to
drop the lock when the time comes, provided that the lock has not
become so contested that threads have started to block in it.

But!  When threads do start to pile up behind this lock, things get
really bad. Now, when the lock is released, there is always one or
more threads in the wait queue, one of which need to be woken.  In a
_heavily_ instrumented kernel (again, nested VM), the act of dropping
the lock and waking/scheduling the next thread to run takes on average
something like 10uSec, or about 33 times as long as the work actually
done in the lock.

Provided that the work done inside of the lock is well bounded, it
would be better to use a spinlock here instead of a blocking mutex.
There seem to be two reasons that this lock is currently a blocking
mutex, however.

1) In the past, it seems like the scanner threads would hold this lock
   for relatively long periods of time where it would not be
   appropriate to use a spin lock.
2) While holding this lock, various threads occasionally need to
   signal a kernel Event as the queue sizes change in order to trigger
   LRU scanning and aging behaviors.  One cannot signal an event while
   holding a spinlock, since it represents an action which might
   result in immediate preemption if the signaled thread ends up
   displacing the signaling thread in the scheduler's run queue.

Issue #1 does not seem to be much of an issue anymore.  The scanner
threads now attempt to batch their work in ways which mean that there
is a worst case amount of time that they might spend in the lock
fetching a batch of work to scan.

Issue #2 was addressed using a bit of static-analysis trickery.
After converting the lock from a critical mutex to a spinlock, the
locations where the events were being directly signaled were replaced
with code which simply records that the need to signal the events are
now "pending".

A special zero-length static analysis capability (the
PendingSignalToken) was added.  This object behaves like a lock from a
static analysis perspective, but under the hood, it actually does
nothing at all.  It is also required to be acquired before the
object's main spinlock

Next annotations were added to the new "PendSignal" method which
demand that the PendingSignalToken needs to be held in order to pend a
signal.

Finally, an RAII style guard was introduced which "acquires" and
"releases" the token, and always flushes out any pending signals when
the guard goes out of scope.  The Acquire/Release methods of the token
itself were private, and the Guard was made a friend of the Token.

So - With all of this in place:

1) No events will ever be signaled from inside of the (now) spinlock.
2) Instead, they will be pended, but this requires holding the token.
3) The token cannot be simply acquired by random code.  It may only be
   acquired by instantiating the guard.
4) A guard can only be instantiated before acquiring the spinlock
   (which is also only ever acquired using a guard object).  This
   means that the token-guard will always go out of scope after the
   lock guard.
5) The guard makes sure to always flush the pending signals, so it is
   now extremely difficult to make a mistake and forget to flush any
   pending signals.
6) The static-annotations make sure that all code patch which _might_
   result in a pended signal will make sure to flush the pending
   signals at a point after the spinlock has been dropped.

The result is a significant performance improvement in scenarios where
the PageQueues lock had been previously heavily contested.

Change-Id: I4c2b26007a58f1771cf8c4b130e18c47893c790f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762186
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: John Grossman <johngro@google.com>