[emu] Shift --edit code to run after staging
This CL moves the code path for editing the emulator configuration to
run after the staging step. This ensures that the paths in the editable
configuration are the staged files rather than the source files, and
that steps like port mapping have already occurred when the
configuration is presented to the user for editing.

Change-Id: Ie18f65284e5f19add3b70fb4d3900e4fba0be073
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773547
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>