[ffx][fho] Add writer infrastructure
This adds a Writer associated type to FfxMain and uses it as an
additional argument to the main() function. Originally we intended this
to be something that would be included in the FfxTool struct, but with
the goal of minimizing complicated macro machinery it seemed better to
move it out to its own thing.

A new trait is introduced, `FfxToolIo`, that currently only mimics the
`ffx_plugin` functions around machine output, and defaults them to
saying no machine output. A followup change will add a concrete type to
the ffx_writer crate that requires a concrete machine type and will impl
this trait with 'better' definitions for those things.

Bug: 114369
Change-Id: Idfc90cab08570e1c2c86ea0381ebaa126c6d238d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/774662
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>