[starnix][ext4] Use exclamation mark as separator for symbolization
The VMO name for ELF segments is used in the symbolizer output to
identify modules. The symbolizer format does not support using ':' or
'}' in the name field and does not have an escape mechanism:

https://fuchsia.dev/fuchsia-src/reference/kernel/symbolizer_markup

This uses 'ext!' instead of 'ext4:' as a prefix to work around this
limitation. This lets tools that parse symbolizer output process these
modules correctly.

Change-Id: I551500ed3e55e00875a226420feb3373794200a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/775622
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>