[fs_management] Allow multiple guids/formats for PartitionMatcher
In cases where we might want to match against either a legacy GUID or a
newer GUID, and we don't know which one ahead of time, it is helpful to
extend the PartitionMatcher to take in a vector of type and instance
guids. Same goes for formats. This change allows for that flexibility.

Testing: fx test open_partition_test-component
Change-Id: I64b1a6d9f6b923c1efbc171f0a58df888ab280ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773673
Reviewed-by: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Nikita Jindal <nikitajindal@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>