[ffx] Build ffx-scrutiny exe and use in builds
This makes the scrutiny binary build in parallel as a separate binary
that can be invoked from the build, and modifies all the places we
currently have it doing so through ffx to use it.

This is not exactly how I wanted to solve this, but it works for now and
a better rearchitecting of how these components interact is for another
commit. It turns out to be annoyingly hard to improve these bits.

A weird side effect of the way I did this (generating fho handlers for
suites in the jinja template) is that ffx has to be labelled a subcommand.
This doesn't hurt anything, it's just an extra impl on the struct, but
it saved trying to hack in a way to tell if the suite was a subcommand
suite or not.

Fixed: 116716
Bug: 116712
Change-Id: I071d5a24bea34a5301c433487a766f1e466204d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771667
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>