[dfv2][compat] UnbindAndRelease waits for Init
It is an error to call Unbind on a driver before it has finished
initializing, so move this check into UnbindAndRelease so it is
not possible to do this incorrectly.

Also make the root device of the compat driver have the correct init
state when being created.

Bug: 117387
Change-Id: I6eb0309416c0857dcf15de34138c067796186c10
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/776163
Reviewed-by: Ian Fisher <iafisher@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>