[fidl][hlcpp] unknown interactions fp::result
HLCPP result unions added support for bidirectionally converting with a
corresponding fpromise::result, however this support was not extended to
unknown interactions. Conversion from fpromise::result to the result
union would work correctly, however, the conversion from result union to
fpromise::result did not account for the transport_err variant in a way
that could cause either a crash or UB when converting from a result
union to fpromise::result.

This change updates the support for conversion with fpromise::result by
changing the type used in the error variant:

- If the method is strict, the result union converts with
  fpromise::result<V, E>, as it would previously.
- If the method is flexible without error syntax, the result union
  converts with fpromise::result<V, fidl::TransportErr>
- If the method is flexible with error syntax, the result union converts
  with fpromise::result<V, std::variant<E, fidl::TransportErr>>.

Bug: 88366
Change-Id: I0bf243c60707b7203d9f171691820ee4318d9be8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/776282
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Commit-Queue: Zachary Stewart <zstewart@google.com>