[fs_management] control component launch type
fshost has a collection and several static filesystem children,
depending on configuration. It needs to be able to specify whether to
use the static child or launch the filesystem instance in the
collection, since the static children are only used for specific
instances.

Add a ComponentType configuration enum which lets users specify which
type of child to launch. Make the default to be dynamic launching in a
collection named fs-collection, which is what most users want.

Test: fx test fs-management-rust-tests
Change-Id: I55bbe71712f5fff6938bc2ee06e7d23b1be9a8c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773685
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>