[audio] fuchsia_audio_device::ControlCreator
This CL implements fuchsia_audio_device::ControlCreator. It
also adds Device class support for dynamic changes to gain
and plug states as well as the SetGain method.

This CL also introduces the concept that a `const` view of
a Device will be the limited subset of commands available to
device observers. As a temporary measure (until a future CL
that introduces the ControlNotify interface), the method
Device::SetControl is used to reserve a device so that it
can be controlled only by that client. SetGain is an example
of a method that can only be called when the device is
controlled.

Bug: 87650

Multiply: audio_services_device_registry_unittests
Multiply: audio_services_device_registry_warning_unittests

Change-Id: I25527b71380d91bcbb7d662dbcff9fc48d8e02f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745606
Reviewed-by: Tom Bergan <tombergan@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>