[ffx][emu] Move host tool hack to the only user
The issue referenced in the get_host_tool() function has been closed for
a very long time, but it did not completely resolve the issue (see
fxr/774028 and its revert fxr/776022) because ffx emu tests out of tree
still rely on this behaviour.

Rather than leave it in or remove it outright, move this over to ffx emu
where it's still used and tie it to a still-current related issue.

Bug: 99321
Change-Id: I912ac785856e6925575f13ae50b8abbc6a6cfea2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/776329
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Daniel Dougherty <doughertyda@google.com>