[fidl][rust] Large message decode support
This implements the "happy path" for the large message decode support in
Rust. It does so by teaching the bindings to recognize both the
prototype (no `LargeMessageInfo` struct) and final RFC (with the struct)
designs. Once the encode logic is updated to only encode the RFC way,
the support for the old decoding method will be removed, and "Bad" sad
path tests will be added as well.

Test: fx test server-suite-rust
Bug: 114259
Change-Id: I516a5ed3657d7d838321121733cd4ce7bfe2ec74
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758953
Reviewed-by: Mitchell Kember <mkember@google.com>
Fuchsia-Auto-Submit: Alex Zaslavsky <azaslavsky@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>